+++ src/components/Table/TableBody/TableBody.js
@@ -1,11 +1,15 @@
 import React, { useContext, useEffect, useRef } from 'react'
 import propTypes from 'prop-types'
 import classNames from 'classnames'
+import { ReactComponent as NoResultsIcon } from '../../../assets/svg/NoResults.svg'
+import languageService from '../../../services/language'
 import { InfiniteList } from '../../../index'
 import TableContext from '../TableContext'
 import { TableRow } from './TableRow'
 import { useTableData } from '../UseTableData'
 import styles from './TableBody.module.scss'
+
+const getTranslation = path => languageService.getTranslation(`${path}`)
 
 const TableBody = ({
   data,
@@ -88,9 +92,21 @@ const TableBody = ({
     ))
   }
 
+  const renderNoResults = () => (
+    <div className={styles.noResults}>
+      <NoResultsIcon />
+      <div className={styles.noResultsTitle}>
+        {getTranslation('noResultsFound')}
+      </div>
+      <div className={styles.noResultsMessage}>
+        {getTranslation('noResultsMessage')}
+      </div>
+    </div>
+  )
+
   const classes = classNames(styles.tableBody, className)
 
-  return (
+  return state.totalItems ? (
     <div className={classes} {...otherProps}>
       <InfiniteList
         rowHeight={rowHeight}
@@ -101,8 +117,10 @@ const TableBody = ({
         isLoading={isLoading}
         loadMoreItems={loadMoreData}
         ref={listRef}
-      ></InfiniteList>
+      />
     </div>
+  ) : (
+    renderNoResults()
   )
 }